Code clean-up: move intercon-related stuff to their own file, and move initialization-related stuff to main.
