+++ entry.S
@@ -2,4 +2,4 @@
 .globl _start
 
 _start:
-	B hack_entry_point
+	B main
