+++ utils.c
@@ -184,35 +184,6 @@ void start_debug_mode() {
 	beep();
 }
 
-int send_to_intercom(int message, int parm) {
-	int result, length = 1;
-
-	switch (message) {
-	case IC_SET_AE:
-		status.ignore_msg = IC_SETTINGS_0;
-		break;
-	case IC_SET_AV_VAL:
-	case IC_SET_TV_VAL:
-		status.ignore_msg = message;
-		break;
-	case IC_RELEASE:
-	case IC_SET_REALTIME_ISO_0:
-	case IC_SET_REALTIME_ISO_1:
-		length = 0;
-		break;
-	case IC_SET_ISO:
-	case IC_SET_AF_POINT:
-	case IC_SET_COLOR_TEMP:
-		length = 2;
-		break;
-	}
-
-	result = SendToIntercom(message, length, parm);
-	SleepTask(INTERCOM_WAIT);
-
-	return result;
-}
-
 #if 0
 // this is a disassembled version of eventproc_release()
 int shutter_release_disasm() {
